now tests fail

+++ py_bind/optv/orientation.pyx
@@ -263,7 +263,7 @@ def full_calibration(Calibration cal,
     cdef:
         vec3d *ref_coord
         np.ndarray[ndim=2, dtype=pos_t] ret
-        np.ndarray[ndim=1, dtype=np.int32_t] used
+        np.ndarray[ndim=1, dtype=np.int64_t] used
         np.ndarray[ndim=1, dtype=pos_t] err_est
         orient_par *orip
         double *residuals
@@ -298,7 +298,7 @@ def full_calibration(Calibration cal,
         raise ValueError("Orientation iteration failed, need better setup.")
     
     ret = np.empty((len(img_pts), 2))
-    used = np.empty(len(img_pts), dtype=np.int_)
+    used = np.empty(len(img_pts), dtype=np.int64)
     
     for ix in range(len(img_pts)):
         ret[ix] = (residuals[2*ix], residuals[2*ix + 1])
